requirements.txt
selenium
undetected_chromedriver
python-dotenv
setuptools
tqdm